backend/requirements.txt
django
djangorestframework
django-filter
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
pytest
pytest-django
python-dotenv
